requirements.txt
flask==2.3.3
tensorflow==2.17.1
pillow==10.2.0
numpy==1.26.3